+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "obsidianki"
-version = "0.3.0"
+version = "0.4.0"
 description = "Generate flashcards from Obsidian notes using AI and sync them to Anki"
 authors = [{name = "ccmdi", email = "ccmdi.dev@gmail.com"}]
 license = {text = "MIT"}
@@ -31,4 +31,4 @@ Repository = "https://github.com/ccmdi/obsidianki"
 Issues = "https://github.com/ccmdi/obsidianki/issues"
 
 [tool.setuptools]
-py-modules = ["main", "obsidian", "anki", "config", "wizard", "cli_handlers"]
+py-modules = ["main"]
